añadido 'obtenerNombrePorId' a los servicios y retoque de la vista del resumen

+++ Modulo-Comun/src/main/java/grupo/a/modulocomun/Repositorios/EntidadBancariaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EntidadBancariaRepository extends JpaRepository<EntidadBancaria, Integer> {
+public interface EntidadBancariaRepository extends JpaRepository<EntidadBancaria, Long> {
 }
