+++ Modulo-Comun/src/main/java/grupo/a/modulocomun/Repositorios/EspecialidadesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EspecialidadesRepository extends JpaRepository<Especialidades, Integer> {
+public interface EspecialidadesRepository extends JpaRepository<Especialidades, Long> {
 }
