+++ Modulo-Comun/src/main/java/grupo/a/modulocomun/Repositorios/TipoTarjetaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoTarjetaRepository extends JpaRepository<TipoTarjeta, Integer> {
+public interface TipoTarjetaRepository extends JpaRepository<TipoTarjeta, Long> {
 }
